withUpdateEmailAddress in the plugins hoc
